checking live websites using getcode()
